functional_testing/minio/requirements.txt
nose
nose-testconfig
nose-parameterized
minio